Rename build -> deserialize
